pyapps/telegram_bot/requirements.txt
# requirements for pybot
python-telegram-bot
pyyaml
icecream